refactor: remove priority_objects, use as an object property
