+++ clovers-front/src/Objects/FlipFace.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from "react";
 import { DeleteButton } from "../DeleteButton";
-import { TextInput } from "../Input";
+import { CheckboxInput, TextInput } from "../Input";
 import { ObjectForm } from "./SceneObject";
 
 export type FlipFace = {
@@ -23,6 +23,12 @@ export const FlipFaceForm = ({
       <DeleteButton path={path} setState={setState} />
       <TextInput
         fieldname="comment"
+        object={object}
+        path={path}
+        setState={setState}
+      />
+      <CheckboxInput
+        fieldname="priority"
         object={object}
         path={path}
         setState={setState}
