+++ clovers-front/src/Objects/Translate.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from "react";
 import { DeleteButton } from "../DeleteButton";
-import { TextInput, TripleNumberInput } from "../Input";
+import { CheckboxInput, TextInput, TripleNumberInput } from "../Input";
 import { ObjectForm, SceneObject } from "./SceneObject";
 
 export type Translate = {
@@ -30,6 +30,12 @@ export const TranslateForm = ({
       />
       <TripleNumberInput
         fieldname="offset"
+        object={object}
+        path={path}
+        setState={setState}
+      />
+      <CheckboxInput
+        fieldname="priority"
         object={object}
         path={path}
         setState={setState}
